Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main()
 {
-    int a;
-    int b;
-    int c;
-    int average;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    float a;
+    float b;
+    float c;
+    float average;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
     average=(a+b+c)/3;
     printf("Average: %.2f",average);
     return 0;
